requirements-dev.txt
aws-cdk-lib==2.137.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.0,<2.0.0
black==24.4.2
ruff==0.5.5
pytest==8.2.2
requests==2.32.3
